chat-personality-analyzer/requirements.txt
streamlit>=1.39.0
python-dotenv>=1.0.1
Pillow>=10.3.0
google-generativeai>=0.3.2
